requirements.txt
asgiref==3.2.10
certifi==2021.5.30
chardet==4.0.0
Django==3.1
django-appconf==1.0.4
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-taggit==1.4.0
djangorestframework==3.12.4
idna==2.10
pilkit==2.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
